chore: hoist and consolidate lib
