+++ components/custom-test/survey-test/graphs/AffordFairhold.tsx
@@ -1,8 +1,8 @@
 import React from "react"
 import SurveyGraphCard from "@/components/custom-test/survey-test/SurveyGraphCard";
 import { PieChart, Pie, Legend, ResponsiveContainer, Cell } from "recharts";
-import { useSurveyContext } from "../../../../app/survey/context";
-import { AFFORD_FAIRHOLD } from "../../../../app/survey/constants";
+import { useSurveyContext } from "@context/surveyContext";
+import { AFFORD_FAIRHOLD } from "../../../../lib/survey/constants";
 
 export const AffordFairhold = () => {
     let { affordFairhold } = useSurveyContext().barOrPie;
