+++ lib/calculator/exceptions.ts
