+++ lib/calculator/format.ts
